moved testObservability tests to separate folder
